reorg part 2
