+++ lib/p5.js
@@ -2954,20 +2954,6 @@ amdclean['p5Image'] = function (require, core, filters) {
   };
   return p5.Image;
 }({}, amdclean['core'], amdclean['filters']);
-amdclean['p5File'] = function (require, core) {
-  var p5 = core;
-  p5.File = function (file, pInst) {
-    this.file = file;
-    this._pInst = pInst;
-    var typeList = file.type.split('/');
-    this.type = typeList[0];
-    this.subtype = typeList[1];
-    this.name = file.name;
-    this.size = file.size;
-    this.data = undefined;
-  };
-  return p5.File;
-}({}, amdclean['core']);
 amdclean['polargeometry'] = function (require) {
   return {
     degreesToRadians: function (x) {
@@ -4068,6 +4054,16 @@ amdclean['environment'] = function (require, core, constants) {
   p5.prototype._frameRate = 0;
   p5.prototype._lastFrameTime = new Date().getTime();
   p5.prototype._targetFrameRate = 60;
+  if (window.console && console.log) {
+    p5.prototype.print = function (args) {
+      var newArgs = JSON.parse(JSON.stringify(args));
+      console.log(newArgs);
+    };
+  } else {
+    p5.prototype.print = function () {
+    };
+  }
+  p5.prototype.println = p5.prototype.print;
   p5.prototype.frameCount = 0;
   p5.prototype.focused = true;
   p5.prototype.cursor = function (type, x, y) {
@@ -4205,6 +4201,116 @@ amdclean['imageimage'] = function (require, core, constants) {
   p5.prototype._tint = null;
   p5.prototype.createImage = function (width, height) {
     return new p5.Image(width, height);
+  };
+  var frames = [];
+  p5.prototype.saveCanvas = function (_cnv, filename, extension) {
+    if (!extension) {
+      extension = p5.prototype._checkFileExtension(filename, extension)[1];
+      if (extension === '') {
+        extension = 'png';
+      }
+    }
+    var cnv;
+    if (_cnv) {
+      cnv = _cnv;
+    } else if (this._curElement && this._curElement.elt) {
+      cnv = this._curElement.elt;
+    }
+    if (p5.prototype._isSafari()) {
+      var aText = 'Hello, Safari user!\n';
+      aText += 'Now capturing a screenshot...\n';
+      aText += 'To save this image,\n';
+      aText += 'go to File --> Save As.\n';
+      alert(aText);
+      window.location.href = cnv.toDataURL();
+    } else {
+      var mimeType;
+      if (typeof extension === 'undefined') {
+        extension = 'png';
+        mimeType = 'image/png';
+      } else {
+        switch (extension) {
+        case 'png':
+          mimeType = 'image/png';
+          break;
+        case 'jpeg':
+          mimeType = 'image/jpeg';
+          break;
+        case 'jpg':
+          mimeType = 'image/jpeg';
+          break;
+        default:
+          mimeType = 'image/png';
+          break;
+        }
+      }
+      var downloadMime = 'image/octet-stream';
+      var imageData = cnv.toDataURL(mimeType);
+      imageData = imageData.replace(mimeType, downloadMime);
+      p5.prototype.downloadFile(imageData, filename, extension);
+    }
+  };
+  p5.prototype.saveFrames = function (fName, ext, _duration, _fps, callback) {
+    var duration = _duration || 3;
+    duration = p5.prototype.constrain(duration, 0, 15);
+    duration = duration * 1000;
+    var fps = _fps || 15;
+    fps = p5.prototype.constrain(fps, 0, 22);
+    var count = 0;
+    var makeFrame = p5.prototype._makeFrame;
+    var cnv = this._curElement.elt;
+    var frameFactory = setInterval(function () {
+        makeFrame(fName + count, ext, cnv);
+        count++;
+      }, 1000 / fps);
+    setTimeout(function () {
+      clearInterval(frameFactory);
+      if (callback) {
+        callback(frames);
+      } else {
+        for (var i = 0; i < frames.length; i++) {
+          var f = frames[i];
+          p5.prototype.downloadFile(f.imageData, f.filename, f.ext);
+        }
+      }
+      frames = [];
+    }, duration + 0.01);
+  };
+  p5.prototype._makeFrame = function (filename, extension, _cnv) {
+    var cnv;
+    if (this) {
+      cnv = this._curElement.elt;
+    } else {
+      cnv = _cnv;
+    }
+    var mimeType;
+    if (!extension) {
+      extension = 'png';
+      mimeType = 'image/png';
+    } else {
+      switch (extension.toLowerCase()) {
+      case 'png':
+        mimeType = 'image/png';
+        break;
+      case 'jpeg':
+        mimeType = 'image/jpeg';
+        break;
+      case 'jpg':
+        mimeType = 'image/jpeg';
+        break;
+      default:
+        mimeType = 'image/png';
+        break;
+      }
+    }
+    var downloadMime = 'image/octet-stream';
+    var imageData = cnv.toDataURL(mimeType);
+    imageData = imageData.replace(mimeType, downloadMime);
+    var thisFrame = {};
+    thisFrame.imageData = imageData;
+    thisFrame.filename = filename;
+    thisFrame.ext = extension;
+    frames.push(thisFrame);
   };
   return p5;
 }({}, amdclean['core'], amdclean['constants']);
@@ -5199,6 +5305,258 @@ amdclean['inputfiles'] = function (require, core, reqwest) {
       }
     });
   };
+  window.URL = window.URL || window.webkitURL;
+  p5.prototype._pWriters = [];
+  p5.prototype.beginRaw = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.beginRecord = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.createOutput = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.createWriter = function (name, extension) {
+    var newPW;
+    for (var i in p5.prototype._pWriters) {
+      if (p5.prototype._pWriters[i].name === name) {
+        newPW = new p5.PrintWriter(name + window.millis(), extension);
+        p5.prototype._pWriters.push(newPW);
+        return newPW;
+      }
+    }
+    newPW = new p5.PrintWriter(name, extension);
+    p5.prototype._pWriters.push(newPW);
+    return newPW;
+  };
+  p5.prototype.endRaw = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.endRecord = function () {
+    throw 'not yet implemented';
+  };
+  p5.PrintWriter = function (filename, extension) {
+    var self = this;
+    this.name = filename;
+    this.content = '';
+    this.print = function (data) {
+      this.content += data;
+    };
+    this.println = function (data) {
+      this.content += data + '\n';
+    };
+    this.flush = function () {
+      this.content = '';
+    };
+    this.close = function () {
+      var arr = [];
+      arr.push(this.content);
+      p5.prototype.writeFile(arr, filename, extension);
+      for (var i in p5.prototype._pWriters) {
+        if (p5.prototype._pWriters[i].name === this.name) {
+          p5.prototype._pWriters.splice(i, 1);
+        }
+      }
+      self.flush();
+      self = {};
+    };
+  };
+  p5.prototype.saveBytes = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.save = function (object, _filename, _options) {
+    var args = arguments;
+    var cnv = this._curElement.elt;
+    if (args.length === 0) {
+      p5.prototype.saveCanvas(cnv);
+      return;
+    } else if (args[0] instanceof p5.Graphics) {
+      p5.prototype.saveCanvas(args[0].elt, args[1], args[2]);
+      return;
+    } else if (args.length === 1 && typeof args[0] === 'string') {
+      p5.prototype.saveCanvas(cnv, args[0]);
+    } else {
+      var extension = _checkFileExtension(args[1], args[2])[1];
+      switch (extension) {
+      case 'json':
+        p5.prototype.saveJSON(args[0], args[1], args[2]);
+        return;
+      case 'txt':
+        p5.prototype.saveStrings(args[0], args[1], args[2]);
+        return;
+      default:
+        if (args[0] instanceof Array) {
+          p5.prototype.saveStrings(args[0], args[1], args[2]);
+        } else if (args[0] instanceof p5.Table) {
+          p5.prototype.saveTable(args[0], args[1], args[2], args[3]);
+        } else if (args[0] instanceof p5.Image) {
+          p5.prototype.saveCanvas(args[0].canvas, args[1]);
+        } else if (args[0] instanceof p5.SoundFile) {
+          p5.prototype.saveSound(args[0], args[1], args[2], args[3]);
+        }
+      }
+    }
+  };
+  p5.prototype.saveJSON = function (json, filename, opt) {
+    var stringify;
+    if (opt) {
+      stringify = JSON.stringify(json);
+    } else {
+      stringify = JSON.stringify(json, undefined, 2);
+    }
+    console.log(stringify);
+    this.saveStrings(stringify.split('\n'), filename, 'json');
+  };
+  p5.prototype.saveJSONObject = p5.prototype.saveJSON;
+  p5.prototype.saveJSONArray = p5.prototype.saveJSON;
+  p5.prototype.saveStream = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.saveStrings = function (list, filename, extension) {
+    var ext = extension || 'txt';
+    var pWriter = this.createWriter(filename, ext);
+    for (var i in list) {
+      if (i < list.length - 1) {
+        pWriter.println(list[i]);
+      } else {
+        pWriter.print(list[i]);
+      }
+    }
+    pWriter.close();
+    pWriter.flush();
+  };
+  p5.prototype.saveXML = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.selectOutput = function () {
+    throw 'not yet implemented';
+  };
+  p5.prototype.saveTable = function (table, filename, options) {
+    var pWriter = this.createWriter(filename, options);
+    var header = table.columns;
+    var sep = ',';
+    if (options === 'tsv') {
+      sep = '\t';
+    }
+    if (options !== 'html') {
+      if (header[0] !== '0') {
+        for (var h = 0; h < header.length; h++) {
+          if (h < header.length - 1) {
+            pWriter.print(header[h] + sep);
+          } else {
+            pWriter.println(header[h]);
+          }
+        }
+      }
+      for (var i = 0; i < table.rows.length; i++) {
+        var j;
+        for (j = 0; j < table.rows[i].arr.length; j++) {
+          if (j < table.rows[i].arr.length - 1) {
+            pWriter.print(table.rows[i].arr[j] + sep);
+          } else if (i < table.rows.length - 1) {
+            pWriter.println(table.rows[i].arr[j]);
+          } else {
+            pWriter.print(table.rows[i].arr[j]);
+          }
+        }
+      }
+    } else {
+      pWriter.println('<html>');
+      pWriter.println('<head>');
+      var str = '  <meta http-equiv="content-type" content';
+      str += '="text/html;charset=utf-8" />';
+      pWriter.println(str);
+      pWriter.println('</head>');
+      pWriter.println('<body>');
+      pWriter.println('  <table>');
+      if (header[0] !== '0') {
+        pWriter.println('    <tr>');
+        for (var k = 0; k < header.length; k++) {
+          var e = escapeHelper(header[k]);
+          pWriter.println('      <td>' + e);
+          pWriter.println('      </td>');
+        }
+        pWriter.println('    </tr>');
+      }
+      for (var row = 0; row < table.rows.length; row++) {
+        pWriter.println('    <tr>');
+        for (var col = 0; col < table.columns.length; col++) {
+          var entry = table.rows[row].getString(col);
+          var htmlEntry = escapeHelper(entry);
+          pWriter.println('      <td>' + htmlEntry);
+          pWriter.println('      </td>');
+        }
+        pWriter.println('    </tr>');
+      }
+      pWriter.println('  </table>');
+      pWriter.println('</body>');
+      pWriter.print('</html>');
+    }
+    pWriter.close();
+    pWriter.flush();
+  };
+  var escapeHelper = function (content) {
+    return content.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/'/g, '&#039;');
+  };
+  p5.prototype.writeFile = function (dataToDownload, filename, extension) {
+    var type = 'application/octet-stream';
+    if (p5.prototype._isSafari()) {
+      type = 'text/plain';
+    }
+    var blob = new Blob(dataToDownload, { 'type': type });
+    var href = window.URL.createObjectURL(blob);
+    p5.prototype.downloadFile(href, filename, extension);
+  };
+  p5.prototype.downloadFile = function (href, fName, extension) {
+    var fx = _checkFileExtension(fName, extension);
+    var filename = fx[0];
+    var ext = fx[1];
+    var a = document.createElement('a');
+    a.href = href;
+    a.download = filename;
+    a.onclick = destroyClickedElement;
+    a.style.display = 'none';
+    document.body.appendChild(a);
+    if (p5.prototype._isSafari()) {
+      var aText = 'Hello, Safari user! To download this file...\n';
+      aText += '1. Go to File --> Save As.\n';
+      aText += '2. Choose "Page Source" as the Format.\n';
+      aText += '3. Name it with this extension: ."' + ext + '"';
+      alert(aText);
+    }
+    a.click();
+    href = null;
+  };
+  function _checkFileExtension(filename, extension) {
+    if (!extension || extension === true || extension === 'true') {
+      extension = '';
+    }
+    if (!filename) {
+      filename = 'untitled';
+    }
+    var ext = '';
+    if (filename && filename.indexOf('.') > -1) {
+      ext = filename.split('.').pop();
+    }
+    if (extension) {
+      if (ext !== extension) {
+        ext = extension;
+        filename = filename + '.' + ext;
+      }
+    }
+    return [
+      filename,
+      ext
+    ];
+  }
+  p5.prototype._checkFileExtension = _checkFileExtension;
+  p5.prototype._isSafari = function () {
+    var x = Object.prototype.toString.call(window.HTMLElement);
+    return x.indexOf('Constructor') > 0;
+  };
+  function destroyClickedElement(event) {
+    document.body.removeChild(event.target);
+  }
   return p5;
 }({}, amdclean['core'], amdclean['reqwest']);
 amdclean['inputkeyboard'] = function (require, core) {
@@ -5920,393 +6278,6 @@ amdclean['mathtrigonometry'] = function (require, core, polargeometry, constants
   };
   return p5;
 }({}, amdclean['core'], amdclean['polargeometry'], amdclean['constants']);
-amdclean['outputfiles'] = function (require, core) {
-  'use strict';
-  var p5 = core;
-  window.URL = window.URL || window.webkitURL;
-  p5.prototype._pWriters = [];
-  p5.prototype.beginRaw = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.beginRecord = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.createOutput = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.createWriter = function (name, extension) {
-    var newPW;
-    for (var i in p5.prototype._pWriters) {
-      if (p5.prototype._pWriters[i].name === name) {
-        newPW = new p5.PrintWriter(name + window.millis(), extension);
-        p5.prototype._pWriters.push(newPW);
-        return newPW;
-      }
-    }
-    newPW = new p5.PrintWriter(name, extension);
-    p5.prototype._pWriters.push(newPW);
-    return newPW;
-  };
-  p5.prototype.endRaw = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.endRecord = function () {
-    throw 'not yet implemented';
-  };
-  p5.PrintWriter = function (filename, extension) {
-    var self = this;
-    this.name = filename;
-    this.content = '';
-    this.print = function (data) {
-      this.content += data;
-    };
-    this.println = function (data) {
-      this.content += data + '\n';
-    };
-    this.flush = function () {
-      this.content = '';
-    };
-    this.close = function () {
-      var arr = [];
-      arr.push(this.content);
-      p5.prototype.writeFile(arr, filename, extension);
-      for (var i in p5.prototype._pWriters) {
-        if (p5.prototype._pWriters[i].name === this.name) {
-          p5.prototype._pWriters.splice(i, 1);
-        }
-      }
-      self.flush();
-      self = {};
-    };
-  };
-  p5.prototype.saveBytes = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.save = function (object, _filename, _options) {
-    var args = arguments;
-    var cnv = this._curElement.elt;
-    if (args.length === 0) {
-      p5.prototype.saveCanvas(cnv);
-      return;
-    } else if (args[0] instanceof p5.Graphics) {
-      p5.prototype.saveCanvas(args[0].elt, args[1], args[2]);
-      return;
-    } else if (args.length === 1 && typeof args[0] === 'string') {
-      p5.prototype.saveCanvas(cnv, args[0]);
-    } else {
-      var extension = _checkFileExtension(args[1], args[2])[1];
-      switch (extension) {
-      case 'json':
-        p5.prototype.saveJSON(args[0], args[1], args[2]);
-        return;
-      case 'txt':
-        p5.prototype.saveStrings(args[0], args[1], args[2]);
-        return;
-      default:
-        if (args[0] instanceof Array) {
-          p5.prototype.saveStrings(args[0], args[1], args[2]);
-        } else if (args[0] instanceof p5.Table) {
-          p5.prototype.saveTable(args[0], args[1], args[2], args[3]);
-        } else if (args[0] instanceof p5.Image) {
-          p5.prototype.saveCanvas(args[0].canvas, args[1]);
-        } else if (args[0] instanceof p5.SoundFile) {
-          p5.prototype.saveSound(args[0], args[1], args[2], args[3]);
-        }
-      }
-    }
-  };
-  p5.prototype.saveJSON = function (json, filename, opt) {
-    var stringify;
-    if (opt) {
-      stringify = JSON.stringify(json);
-    } else {
-      stringify = JSON.stringify(json, undefined, 2);
-    }
-    console.log(stringify);
-    this.saveStrings(stringify.split('\n'), filename, 'json');
-  };
-  p5.prototype.saveJSONObject = p5.prototype.saveJSON;
-  p5.prototype.saveJSONArray = p5.prototype.saveJSON;
-  p5.prototype.saveStream = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.saveStrings = function (list, filename, extension) {
-    var ext = extension || 'txt';
-    var pWriter = this.createWriter(filename, ext);
-    for (var i in list) {
-      if (i < list.length - 1) {
-        pWriter.println(list[i]);
-      } else {
-        pWriter.print(list[i]);
-      }
-    }
-    pWriter.close();
-    pWriter.flush();
-  };
-  p5.prototype.saveXML = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.selectOutput = function () {
-    throw 'not yet implemented';
-  };
-  p5.prototype.saveTable = function (table, filename, options) {
-    var pWriter = this.createWriter(filename, options);
-    var header = table.columns;
-    var sep = ',';
-    if (options === 'tsv') {
-      sep = '\t';
-    }
-    if (options !== 'html') {
-      if (header[0] !== '0') {
-        for (var h = 0; h < header.length; h++) {
-          if (h < header.length - 1) {
-            pWriter.print(header[h] + sep);
-          } else {
-            pWriter.println(header[h]);
-          }
-        }
-      }
-      for (var i = 0; i < table.rows.length; i++) {
-        var j;
-        for (j = 0; j < table.rows[i].arr.length; j++) {
-          if (j < table.rows[i].arr.length - 1) {
-            pWriter.print(table.rows[i].arr[j] + sep);
-          } else if (i < table.rows.length - 1) {
-            pWriter.println(table.rows[i].arr[j]);
-          } else {
-            pWriter.print(table.rows[i].arr[j]);
-          }
-        }
-      }
-    } else {
-      pWriter.println('<html>');
-      pWriter.println('<head>');
-      var str = '  <meta http-equiv="content-type" content';
-      str += '="text/html;charset=utf-8" />';
-      pWriter.println(str);
-      pWriter.println('</head>');
-      pWriter.println('<body>');
-      pWriter.println('  <table>');
-      if (header[0] !== '0') {
-        pWriter.println('    <tr>');
-        for (var k = 0; k < header.length; k++) {
-          var e = escapeHelper(header[k]);
-          pWriter.println('      <td>' + e);
-          pWriter.println('      </td>');
-        }
-        pWriter.println('    </tr>');
-      }
-      for (var row = 0; row < table.rows.length; row++) {
-        pWriter.println('    <tr>');
-        for (var col = 0; col < table.columns.length; col++) {
-          var entry = table.rows[row].getString(col);
-          var htmlEntry = escapeHelper(entry);
-          pWriter.println('      <td>' + htmlEntry);
-          pWriter.println('      </td>');
-        }
-        pWriter.println('    </tr>');
-      }
-      pWriter.println('  </table>');
-      pWriter.println('</body>');
-      pWriter.print('</html>');
-    }
-    pWriter.close();
-    pWriter.flush();
-  };
-  var escapeHelper = function (content) {
-    return content.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/'/g, '&#039;');
-  };
-  p5.prototype.writeFile = function (dataToDownload, filename, extension) {
-    var type = 'application/octet-stream';
-    if (p5.prototype._isSafari()) {
-      type = 'text/plain';
-    }
-    var blob = new Blob(dataToDownload, { 'type': type });
-    var href = window.URL.createObjectURL(blob);
-    p5.prototype.downloadFile(href, filename, extension);
-  };
-  p5.prototype.downloadFile = function (href, fName, extension) {
-    var fx = _checkFileExtension(fName, extension);
-    var filename = fx[0];
-    var ext = fx[1];
-    var a = document.createElement('a');
-    a.href = href;
-    a.download = filename;
-    a.onclick = destroyClickedElement;
-    a.style.display = 'none';
-    document.body.appendChild(a);
-    if (p5.prototype._isSafari()) {
-      var aText = 'Hello, Safari user! To download this file...\n';
-      aText += '1. Go to File --> Save As.\n';
-      aText += '2. Choose "Page Source" as the Format.\n';
-      aText += '3. Name it with this extension: ."' + ext + '"';
-      alert(aText);
-    }
-    a.click();
-    href = null;
-  };
-  function _checkFileExtension(filename, extension) {
-    if (!extension || extension === true || extension === 'true') {
-      extension = '';
-    }
-    if (!filename) {
-      filename = 'untitled';
-    }
-    var ext = '';
-    if (filename && filename.indexOf('.') > -1) {
-      ext = filename.split('.').pop();
-    }
-    if (extension) {
-      if (ext !== extension) {
-        ext = extension;
-        filename = filename + '.' + ext;
-      }
-    }
-    return [
-      filename,
-      ext
-    ];
-  }
-  p5.prototype._checkFileExtension = _checkFileExtension;
-  p5.prototype._isSafari = function () {
-    var x = Object.prototype.toString.call(window.HTMLElement);
-    return x.indexOf('Constructor') > 0;
-  };
-  function destroyClickedElement(event) {
-    document.body.removeChild(event.target);
-  }
-  return p5;
-}({}, amdclean['core']);
-amdclean['outputimage'] = function (require, core) {
-  'use strict';
-  var p5 = core;
-  var frames = [];
-  p5.prototype.saveCanvas = function (_cnv, filename, extension) {
-    if (!extension) {
-      extension = p5.prototype._checkFileExtension(filename, extension)[1];
-      if (extension === '') {
-        extension = 'png';
-      }
-    }
-    var cnv;
-    if (_cnv) {
-      cnv = _cnv;
-    } else if (this._curElement && this._curElement.elt) {
-      cnv = this._curElement.elt;
-    }
-    if (p5.prototype._isSafari()) {
-      var aText = 'Hello, Safari user!\n';
-      aText += 'Now capturing a screenshot...\n';
-      aText += 'To save this image,\n';
-      aText += 'go to File --> Save As.\n';
-      alert(aText);
-      window.location.href = cnv.toDataURL();
-    } else {
-      var mimeType;
-      if (typeof extension === 'undefined') {
-        extension = 'png';
-        mimeType = 'image/png';
-      } else {
-        switch (extension) {
-        case 'png':
-          mimeType = 'image/png';
-          break;
-        case 'jpeg':
-          mimeType = 'image/jpeg';
-          break;
-        case 'jpg':
-          mimeType = 'image/jpeg';
-          break;
-        default:
-          mimeType = 'image/png';
-          break;
-        }
-      }
-      var downloadMime = 'image/octet-stream';
-      var imageData = cnv.toDataURL(mimeType);
-      imageData = imageData.replace(mimeType, downloadMime);
-      p5.prototype.downloadFile(imageData, filename, extension);
-    }
-  };
-  p5.prototype.saveFrames = function (fName, ext, _duration, _fps, callback) {
-    var duration = _duration || 3;
-    duration = p5.prototype.constrain(duration, 0, 15);
-    duration = duration * 1000;
-    var fps = _fps || 15;
-    fps = p5.prototype.constrain(fps, 0, 22);
-    var count = 0;
-    var makeFrame = p5.prototype._makeFrame;
-    var cnv = this._curElement.elt;
-    var frameFactory = setInterval(function () {
-        makeFrame(fName + count, ext, cnv);
-        count++;
-      }, 1000 / fps);
-    setTimeout(function () {
-      clearInterval(frameFactory);
-      if (callback) {
-        callback(frames);
-      } else {
-        for (var i = 0; i < frames.length; i++) {
-          var f = frames[i];
-          p5.prototype.downloadFile(f.imageData, f.filename, f.ext);
-        }
-      }
-      frames = [];
-    }, duration + 0.01);
-  };
-  p5.prototype._makeFrame = function (filename, extension, _cnv) {
-    var cnv;
-    if (this) {
-      cnv = this._curElement.elt;
-    } else {
-      cnv = _cnv;
-    }
-    var mimeType;
-    if (!extension) {
-      extension = 'png';
-      mimeType = 'image/png';
-    } else {
-      switch (extension.toLowerCase()) {
-      case 'png':
-        mimeType = 'image/png';
-        break;
-      case 'jpeg':
-        mimeType = 'image/jpeg';
-        break;
-      case 'jpg':
-        mimeType = 'image/jpeg';
-        break;
-      default:
-        mimeType = 'image/png';
-        break;
-      }
-    }
-    var downloadMime = 'image/octet-stream';
-    var imageData = cnv.toDataURL(mimeType);
-    imageData = imageData.replace(mimeType, downloadMime);
-    var thisFrame = {};
-    thisFrame.imageData = imageData;
-    thisFrame.filename = filename;
-    thisFrame.ext = extension;
-    frames.push(thisFrame);
-  };
-  return p5;
-}({}, amdclean['core']);
-amdclean['outputtext_area'] = function (require, core) {
-  'use strict';
-  var p5 = core;
-  if (window.console && console.log) {
-    p5.prototype.print = function (args) {
-      var newArgs = JSON.parse(JSON.stringify(args));
-      console.log(newArgs);
-    };
-  } else {
-    p5.prototype.print = function () {
-    };
-  }
-  p5.prototype.println = p5.prototype.print;
-  return p5;
-}({}, amdclean['core']);
 amdclean['renderingrendering'] = function (require, core, constants, p5Graphics2D, p5Graphics3D) {
   var p5 = core;
   var constants = constants;
@@ -7077,7 +7048,7 @@ amdclean['structure'] = function (require, core) {
   };
   return p5;
 }({}, amdclean['core']);
-amdclean['transform'] = function (require, core, constants, outputtext_area) {
+amdclean['transform'] = function (require, core, constants) {
   'use strict';
   var p5 = core;
   var constants = constants;
@@ -7160,7 +7131,7 @@ amdclean['transform'] = function (require, core, constants, outputtext_area) {
     return this;
   };
   return p5;
-}({}, amdclean['core'], amdclean['constants'], amdclean['outputtext_area']);
+}({}, amdclean['core'], amdclean['constants']);
 amdclean['typographyattributes'] = function (require, core, constants) {
   'use strict';
   var p5 = core;
@@ -7283,7 +7254,7 @@ amdclean['typographyloading_displaying'] = function (require, core, helpers) {
   };
   return p5;
 }({}, amdclean['core'], amdclean['helpers']);
-amdclean['src_app'] = function (require, core, p5Color, p5Element, p5Graphics2D, p5Graphics3D, p5Image, p5File, p5Vector, p5TableRow, p5Table, colorcreating_reading, colorsetting, constants, dataconversion, dataarray_functions, datastring_functions, environment, imageimage, imageloading_displaying, imagepixels, inputfiles, inputkeyboard, inputacceleration, inputmouse, inputtime_date, inputtouch, mathmath, mathcalculation, mathrandom, mathnoise, mathtrigonometry, outputfiles, outputimage, outputtext_area, renderingrendering, shape2d_primitives, shape3d_primitives, shapeattributes, shapecurves, shapevertex, structure, transform, typographyattributes, typographyloading_displaying, shaders) {
+amdclean['src_app'] = function (require, core, p5Color, p5Element, p5Graphics2D, p5Graphics3D, p5Image, p5Vector, p5TableRow, p5Table, colorcreating_reading, colorsetting, constants, dataconversion, dataarray_functions, datastring_functions, environment, imageimage, imageloading_displaying, imagepixels, inputfiles, inputkeyboard, inputacceleration, inputmouse, inputtime_date, inputtouch, mathmath, mathcalculation, mathrandom, mathnoise, mathtrigonometry, renderingrendering, shape2d_primitives, shape3d_primitives, shapeattributes, shapecurves, shapevertex, structure, transform, typographyattributes, typographyloading_displaying, shaders) {
   'use strict';
   var p5 = core;
   var _globalInit = function () {
@@ -7299,6 +7270,6 @@ amdclean['src_app'] = function (require, core, p5Color, p5Element, p5Graphics2D,
     window.addEventListener('load', _globalInit, false);
   }
   return p5;
-}({}, amdclean['core'], amdclean['p5Color'], amdclean['p5Element'], amdclean['p5Graphics2D'], amdclean['p5Graphics3D'], amdclean['p5Image'], amdclean['p5File'], amdclean['p5Vector'], amdclean['p5TableRow'], amdclean['p5Table'], amdclean['colorcreating_reading'], amdclean['colorsetting'], amdclean['constants'], amdclean['dataconversion'], amdclean['dataarray_functions'], amdclean['datastring_functions'], amdclean['environment'], amdclean['imageimage'], amdclean['imageloading_displaying'], amdclean['imagepixels'], amdclean['inputfiles'], amdclean['inputkeyboard'], amdclean['inputacceleration'], amdclean['inputmouse'], amdclean['inputtime_date'], amdclean['inputtouch'], amdclean['mathmath'], amdclean['mathcalculation'], amdclean['mathrandom'], amdclean['mathnoise'], amdclean['mathtrigonometry'], amdclean['outputfiles'], amdclean['outputimage'], amdclean['outputtext_area'], amdclean['renderingrendering'], amdclean['shape2d_primitives'], amdclean['shape3d_primitives'], amdclean['shapeattributes'], amdclean['shapecurves'], amdclean['shapevertex'], amdclean['structure'], amdclean['transform'], amdclean['typographyattributes'], amdclean['typographyloading_displaying'], amdclean['shaders']);
+}({}, amdclean['core'], amdclean['p5Color'], amdclean['p5Element'], amdclean['p5Graphics2D'], amdclean['p5Graphics3D'], amdclean['p5Image'], amdclean['p5Vector'], amdclean['p5TableRow'], amdclean['p5Table'], amdclean['colorcreating_reading'], amdclean['colorsetting'], amdclean['constants'], amdclean['dataconversion'], amdclean['dataarray_functions'], amdclean['datastring_functions'], amdclean['environment'], amdclean['imageimage'], amdclean['imageloading_displaying'], amdclean['imagepixels'], amdclean['inputfiles'], amdclean['inputkeyboard'], amdclean['inputacceleration'], amdclean['inputmouse'], amdclean['inputtime_date'], amdclean['inputtouch'], amdclean['mathmath'], amdclean['mathcalculation'], amdclean['mathrandom'], amdclean['mathnoise'], amdclean['mathtrigonometry'], amdclean['renderingrendering'], amdclean['shape2d_primitives'], amdclean['shape3d_primitives'], amdclean['shapeattributes'], amdclean['shapecurves'], amdclean['shapevertex'], amdclean['structure'], amdclean['transform'], amdclean['typographyattributes'], amdclean['typographyloading_displaying'], amdclean['shaders']);
 return amdclean['src_app'];
 }));
