+++ src/3d/3d_primitives.js
@@ -1,6 +1,4 @@
 /**
- * @module Shape
- * @submodule 3D Primitives
  * @for p5
  * @requires core
  * @requires canvas
