+++ src/events/acceleration.js
@@ -1,5 +1,5 @@
 /**
- * @module Input
+ * @module Events
  * @submodule Acceleration
  * @for p5
  * @requires core
